refactor(context): implement context for state management
